+++ blocks/editor.js
@@ -1,7 +1,6 @@
 
 /* === Radio Station Block Editor Scripts === */
 
-const el = window.wp.element.createElement;
 const { Icon } = wp.components;
 
 /* --- Subscribe to Block State --- */